Use int data type so the platform-specific one can be selected at runtime

+++ calc/calc.go
@@ -2,24 +2,24 @@ package calc
 
 // Adds two numbers.
 // Returns the result.
-func Add(a, b int64) int64 {
+func Add(a, b int) int {
 	return a+b
 }
 
 // Substracts two numbers.
 // Returns the result.
-func Substract(a, b int64) int64 {
+func Substract(a, b int) int {
 	return a-b
 }
 
 // Multiples two numbers.
 // Returns the result.
-func Multiply(a, b int64) int64 {
+func Multiply(a, b int) int {
 	return a*b
 }
 
 // Divides two numbers.
 // Returns the result.
-func Divide(a, b int64) float64 {
+func Divide(a, b int) float64 {
 	return float64(a/b)
 }
