+++ calc_goblin_test/calc_goblin_test.go
@@ -10,15 +10,15 @@ func Test(t *testing.T) {
 	g := Goblin(t)
 	g.Describe("Calculator", func() {
 		g.It("should add two numbers ", func() {
-			g.Assert(Add(1, 2)).Equal(int64(3))
+			g.Assert(Add(1, 2)).Equal(3)
 		})
 
 		g.It("should substract two numbers", func() {
-			g.Assert(Substract(5, 3)).Equal(int64(2))
+			g.Assert(Substract(5, 3)).Equal(2)
 		})
 
 		g.It("should multiply two numbers", func() {
-			g.Assert(Multiply(5, 6)).Equal(int64(30))
+			g.Assert(Multiply(5, 6)).Equal(30)
 		})
 
 		g.It("should divide two numbers", func() {
